Distinguish between decoding a name and a value for a MAP
I thought it might be easier if there was a way to determine if a value was
being used as a key for a conversion function.  While I was in the code, I
added support for a reference table since I know I'm going to be adding that
support Real Soon Now, and I'd rather keep this all coroutine safe.