A quick encode/decode test.
This is to test some MAPs from the RFC.  The RFC listed the order of the
mapping whereas with Lua, we don't get a particular order, but that doesn't
matter to Lua.  So after the RFC test, we run another test with the same
map, this time, just encoding and decoding and check to see if the src data
matches the decoded data, without regard to order.